refactor module variable handling
