+++ feature_engine/_base_transformers/base_numerical.py
@@ -7,7 +7,7 @@ import pandas as pd
 from sklearn.base import BaseEstimator, TransformerMixin
 from sklearn.utils.validation import check_is_fitted
 
-from feature_engine._variable_handling.variable_type_selection import (
+from feature_engine.variable_handling.variable_type_selection import (
     _find_or_check_numerical_variables,
 )
 from feature_engine.dataframe_checks import (
