+++ feature_engine/outliers/artbitrary.py
@@ -24,7 +24,7 @@ from feature_engine._docstrings.methods import (
     _fit_transform_docstring,
 )
 from feature_engine._docstrings.substitute import Substitution
-from feature_engine._variable_handling.variable_type_selection import (
+from feature_engine.variable_handling.variable_type_selection import (
     _find_or_check_numerical_variables,
 )
 from feature_engine.dataframe_checks import (
