+++ feature_engine/selection/drop_features.py
@@ -2,7 +2,7 @@ from typing import List, Union
 
 import pandas as pd
 
-from feature_engine._variable_handling.variable_type_selection import (
+from feature_engine.variable_handling.variable_type_selection import (
     _find_all_variables,
 )
 from feature_engine.dataframe_checks import check_X
