+++ tests/test_variable_handling/test_init_parameter_checks.py
@@ -1,6 +1,6 @@
 import pytest
 
-from feature_engine._variable_handling.init_parameter_checks import (
+from feature_engine.variable_handling._init_parameter_checks import (
     _check_init_parameter_variables,
 )
